FIXED: overflow error in animation timing calculation
git-svn-id: svn://svn.code.sf.net/p/nel/code/trunk@359 bf685bcc-7839-4fcd-9254-36823579bede

+++ snowballs2/client/src/animation.cpp
@@ -276,7 +276,7 @@ void	updateAnimation()
 	}
 
 	// compute new animation position depending of the current time
-	PlayListManager->animate (float(CTime::getLocalTime ())/1000.0f);
+	PlayListManager->animate (double(CTime::getLocalTime ())/1000.0f);
 }
 
 void	releaseAnimation()
